Original Pinboard FireFox extension
This is the original official Pinboard FireFox extension which can be
downloaded here: https://pinboard.in/extensions/firefox/pinboardff.xpi

It has some annoying issues:

1) When using 'Save to Pinboard' the current webpage with the context menu (right-click
mouse) the title form field is always set to undefined

2) The 'Read later on Pinboard' action in the context menu (right-click
mouse) does not work.

Both are caused by not taking into consideration that the user may want
to use the right-click mouse context menu to save the current webpage
instead of a link below the mouse cursor.